Migrate initDB functions to setupDatabase.js to organize/minimize server.js file
